requirements.txt
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
cloudpickle==1.3.0
dash==1.19.0
dash-bootstrap-components==0.12.2
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
gunicorn==20.1.0
networkx==2.5.1
numpy==1.20.3
pandas==1.2.4
PasteDeploy==2.1.0
pickleshare==0.7.5
plotly==4.14.3
prompt-toolkit==3.0.18
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
retrying==1.3.3
sec-edgar-downloader==4.0.3
urllib3==1.26.4